+++ test/MultisigFreezeGuard-MultisigFreezeVoting.test.ts
@@ -175,7 +175,7 @@ describe("Child Multisig DAO with Multisig Parent", () => {
 
     // Deploy MultisigFreezeGuard contract with a 60 block timelock period and 60 block execution period
     const freezeGuardSetupData = abiCoder.encode(
-      ["uint256", "uint256", "address", "address", "address"],
+      ["uint64", "uint64", "address", "address", "address"],
       [
         60, // Timelock period
         60, // Execution period
